Move handleESC as part of Modal.js
