+++ admin_ui/app/org/OrgEditPermissionsModals.js
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect } from 'react';
+import React, { useContext } from 'react';
 import { useObservable } from 'rxjs-hooks';
 import Store from '../Store';
 import { Link } from '@reach/router';
@@ -17,26 +17,6 @@ const OrgEditPermissionsModals = () => {
     () => store.org.currentPermissionsModal$,
     store.org.currentPermissionsModal$.getValue()
   );
-
-  const handleESC = (e) => {
-    if (e.key === "Escape") {
-      // ESC key pressed. Hide any modals
-      store.org.currentPermissionsModal$.next('');
-    }
-  };
-
-  // Hide any modal on ESC keypress
-  // This will be an effect (with cleanup).
-  // We pass [] as second argument to avoid multiple add + remove invocaitons
-  // https://reactjs.org/docs/hooks-effect.html#tip-optimizing-performance-by-skipping-effects for details
-  useEffect(() => {
-    // console.log('useEffect called!!!');
-    window.document.addEventListener('keydown', handleESC);
-    return () => {
-      // console.log('useEffect: removing listener');
-      window.removeEventListener('keydown', handleESC);
-    };
-  }, []);
 
   if (currentModal === 'CREATE') {
     return (
