+++ admin_ui/app/role/RoleDeleteModal.js
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect } from 'react';
+import React, { useContext } from 'react';
 import { useObservable } from 'rxjs-hooks';
 import Store from '../Store';
 import Modal from '../../components/Modal';
@@ -14,26 +14,6 @@ const RoleDeleteModal = () => {
     () => store.role.deleteModal$,
     store.role.deleteModal$.getValue()
   );
-
-  const handleESC = (e) => {
-    if (e.key === 'Escape') {
-      // ESC key pressed. Hide any modals
-      store.role.deleteModal$.next('');
-    }
-  };
-
-  // Hide any modal on ESC keypress
-  // This will be an effect (with cleanup).
-  // We pass [] as second argument to avoid multiple add + remove invocaitons
-  // https://reactjs.org/docs/hooks-effect.html#tip-optimizing-performance-by-skipping-effects for details
-  useEffect(() => {
-    // console.log('useEffect called!!!');
-    window.document.addEventListener('keydown', handleESC);
-    return () => {
-      // console.log('useEffect: removing listener');
-      window.removeEventListener('keydown', handleESC);
-    };
-  }, []);
 
   if (deleteModal === 'DELETE') {
     return (
