+++ admin_ui/components/Modal.js
@@ -1,10 +1,28 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
 function Modal(props) {
-  console.log(props.className);
+  // Invoke the `onClose` (required) prop when ESC key is pressed
+  const handleESC = (e) => {
+    if (e.key === 'Escape') {
+      props.onClose();
+    }
+  };
+
+  // Hide the currently showing modal on ESC keypress
+  // This will be an effect (with cleanup).
+  // https://reactjs.org/docs/hooks-effect.html#tip-optimizing-performance-by-skipping-effects for details
+  useEffect(() => {
+    console.log('useEffect called!!!');
+    window.document.addEventListener('keydown', handleESC);
+    return () => {
+      console.log('useEffect: removing listener');
+      window.removeEventListener('keydown', handleESC);
+    };
+  });
+
   // Use a portal for the modal
   // to attach it as a child of <div id="root">
   return ReactDOM.createPortal(
